improve organization and remove forced moves
