+++ src/test/java/com/dlb/chess/test/unwinnability/TestUnwinnabilityFullWinnable.java
@@ -1,7 +1,6 @@
 package com.dlb.chess.test.unwinnability;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
 
@@ -11,8 +10,6 @@ import org.junit.jupiter.api.Test;
 import com.dlb.chess.board.Board;
 import com.dlb.chess.board.enums.Side;
 import com.dlb.chess.common.Nulls;
-import com.dlb.chess.common.ucimove.utility.UciMoveUtility;
-import com.dlb.chess.model.UciMove;
 import com.dlb.chess.test.RestrictTestConstants;
 import com.dlb.chess.test.common.utility.PgnExtensionUtility;
 import com.dlb.chess.test.model.PgnTestCase;
@@ -23,17 +20,17 @@ import com.dlb.chess.unwinnability.UnwinnabilityFullAnalysis;
 import com.dlb.chess.unwinnability.UnwinnabilityFullVerdict;
 import com.dlb.chess.unwinnability.UnwinnableFullAnalyzer;
 
-class TestUnwinnabilityFullForLichessGamesHavingHelpMate {
+class TestUnwinnabilityFullWinnable {
 
-  private static final Logger logger = Nulls.getLogger(TestUnwinnabilityFullForLichessGamesHavingHelpMate.class);
+  private static final Logger logger = Nulls.getLogger(TestUnwinnabilityFullWinnable.class);
 
   /** Cap on files tested when the smoke restriction is active. */
   private static final int MAX_FILES = 10;
 
   /**
    * For each Lichess game in the helpmate corpus, the analyzer's verdict is {@link UnwinnabilityFullVerdict#WINNABLE}
-   * for the side to move. The companion {@link #mateLinesActuallyCheckmate} test handles the orthogonal claim that
-   * the returned mate line is a valid checkmate sequence.
+   * for the side to move. The companion {@link #mateLinesActuallyCheckmate} test handles the orthogonal claim that the
+   * returned mate line is a valid checkmate sequence.
    */
   @SuppressWarnings("static-method")
   @Test
@@ -49,27 +46,6 @@ class TestUnwinnabilityFullForLichessGamesHavingHelpMate {
     }
   }
 
-  /**
-   * Property check: when the analyzer returns a mate line for a known-winnable position, playing the mate line out
-   * from the start position (after any forced-move preamble) must deliver checkmate against the intended winner's
-   * opponent. Decoupled from {@link #verdictsAreWinnable} so a regression in mate-line construction reports
-   * separately from a regression in verdict computation.
-   */
-  @SuppressWarnings("static-method")
-  @Test
-  void mateLinesActuallyCheckmate() {
-    for (final PgnTestCase testCaseHavingHelpmate : helpmateFixtures()) {
-      logger.info(testCaseHavingHelpmate.pgnName());
-      final PgnTestCase lichessTestCase = PgnTestCaseCatalog
-          .findTestCase(calculateCorrespondingLichessGame(testCaseHavingHelpmate.pgnName()));
-      final Board board = lichessTestCase.finalPosition();
-      final String fen = lichessTestCase.finalFen();
-      final Side winner = board.getHavingMove();
-      final UnwinnabilityFullAnalysis analysis = UnwinnableFullAnalyzer.unwinnableFull(board, winner);
-      assertHelpmateLine(fen, winner, analysis.mateLine());
-    }
-  }
-
   private static List<PgnTestCase> helpmateFixtures() {
     final PgnTestCaseList testCaseHavingHelpmateList = PgnTestCaseCatalog
         .getTestList(PgnTest.CHA_LICHESS_QUICK_DEPTH_ABOVE_FOUR_WINNABLE_FOR_FLAGGING_WITH_HELPMATE);
@@ -77,22 +53,6 @@ class TestUnwinnabilityFullForLichessGamesHavingHelpMate {
       return testCaseHavingHelpmateList.list();
     }
     return testCaseHavingHelpmateList.list().subList(0, Math.min(MAX_FILES, testCaseHavingHelpmateList.list().size()));
-  }
-
-  private static void assertHelpmateLine(String fen, Side winner, List<UciMove> mateLine) {
-    final var board = new Board(fen, false);
-    advanceForcedMoves(board);
-    for (final UciMove uciMove : mateLine) {
-      board.move(UciMoveUtility.convertUciMoveToMoveSpecification(board, uciMove));
-    }
-    assertEquals(winner.getOppositeSide(), board.getHavingMove());
-    assertTrue(board.isCheckmate());
-  }
-
-  private static void advanceForcedMoves(Board board) {
-    while (board.getLegalMoves().size() == 1 && !board.isFivefoldRepetition() && !board.isSeventyFiveMove()) {
-      board.move(Nulls.getFirst(board.getLegalMoves()).moveSpecification());
-    }
   }
 
   private static String calculateCorrespondingLichessGame(String lichessGameHelpmate) {
